hrapi/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
numpy==1.21.1
pandas==1.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
